requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
et-xmlfile==1.0.1
google-api-core==1.14.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.11.2
google-cloud-core==0.29.1
google-cloud-datastore==1.7.3
google-cloud-logging==1.10.0
google-cloud-pubsub==0.45.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpc-google-iam-v1==0.12.3
grpcio==1.19.0
httplib2==0.12.1
idna==2.8
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.1.1
objectpath==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.15.2
wincertstore==0.2